shifting to single css and fixing routing
Put all the css into one file and  then made sure paths were absolute. Also put the logo up on the nav bar. Trying to get footers to behave on housing page, but that doesn't work. It is fine on the giveaid and needaid pages where the main tag is empty (weird)